requirements.txt
dagshub==0.2.19
dataclasses-json==0.5.12
flake8==6.0.0
fqdn==1.5.1
gql==3.4.1
gql-query-builder==0.1.7
ipywidgets==7.7.5
isoduration==20.11.0
jsonpointer==2.4
mlflow==2.5.0
networkx==3.1
numba==0.57.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
pathvalidate==3.0.0
pytest-env==0.8.2
pytest-git==1.7.0
regex==2023.6.3
requests-toolbelt==0.10.1
respx==0.20.1
sympy==1.12
tensorflow==2.13.0
tqdm==4.65.0
treelib==1.6.4
triton==2.0.0
uri-template==1.3.0
webcolors==1.13
dvc==2.58.1
dvc-s3==2.23.0
pycocotools==2.0.6
python-dotenv==1.0.0
